rotate working, Asteroid and Spaceship classes extend Player.

+++ Player.java
@@ -1,17 +1,17 @@
 public class Player{
 
-	private int x;
-	private int y;
+	private double x;
+	private double y;
 
-	public Player(int x, int y){
+	public Player(double x, double y){
 		this.x = x;
 		this.y = y;
 	}
 
-	public int getX(){
+	public double getX(){
 		return x;
 	}
-	public int getY(){
+	public double getY(){
 		return y;
 	}
 }
